refactor: move query and verify to separate file
