+++ crates/sxt-posql-cli/src/args.rs
@@ -1,7 +1,5 @@
-use clap::{Args, Parser, Subcommand};
-use subxt::utils::H256;
-use sxt_proof_of_sql_sdk::SxTClient;
-use sxt_proof_of_sql_sdk_local::CommitmentScheme;
+use crate::query_and_verify::QueryAndVerifySdkArgs;
+use clap::{Parser, Subcommand};
 
 /// Struct to define and parse command-line arguments for Proof of SQL Client.
 ///
@@ -22,94 +20,4 @@ pub struct ProofOfSqlSdkArgs {
 #[derive(Subcommand, Debug, Clone, PartialEq, Eq)]
 pub enum ProofOfSqlSdkSubcommands {
     QueryAndVerify(QueryAndVerifySdkArgs),
-}
-
-#[derive(Args, Debug, Clone, PartialEq, Eq)]
-pub struct QueryAndVerifySdkArgs {
-    /// Root URL for the Prover service
-    ///
-    /// This URL is used for interacting with the prover service.
-    /// Can be set via PROVER_ROOT_URL environment variable.
-    #[arg(
-        long,
-        value_name = "PROVER_ROOT_URL",
-        default_value = "https://api.makeinfinite.dev",
-        env = "PROVER_ROOT_URL"
-    )]
-    pub prover_root_url: String,
-
-    /// Root URL for the Auth service
-    ///
-    /// Used for authentication requests.
-    /// Can be set via AUTH_ROOT_URL environment variable.
-    #[arg(
-        long,
-        value_name = "AUTH_ROOT_URL",
-        default_value = "https://proxy.api.makeinfinite.dev",
-        env = "AUTH_ROOT_URL"
-    )]
-    pub auth_root_url: String,
-
-    /// URL for the Substrate node service
-    ///
-    /// Specifies the Substrate node endpoint used for accessing commitment data.
-    /// Can be set via SUBSTRATE_NODE_URL environment variable.
-    #[arg(
-        long,
-        value_name = "SUBSTRATE_NODE_URL",
-        default_value = "wss://rpc.testnet.sxt.network",
-        env = "SUBSTRATE_NODE_URL"
-    )]
-    pub substrate_node_url: String,
-
-    /// API Key for Space and Time (SxT) services
-    ///
-    /// The API key required for authorization with Space and Time services.
-    /// Can be set via SXT_API_KEY environment variable.
-    #[arg(long, value_name = "SXT_API_KEY", env = "SXT_API_KEY")]
-    pub sxt_api_key: String,
-
-    /// SQL query to execute and verify
-    #[arg(short, long, value_name = "QUERY", help = "SQL query to run")]
-    pub query: String,
-
-    /// SxT chain block hash to perform the query at.
-    #[arg(long)]
-    pub block_hash: Option<H256>,
-
-    /// Commitment scheme to use for the query
-    #[arg(
-        long,
-        value_enum,
-        env,
-        default_value_t = CommitmentScheme::DynamicDory,
-    )]
-    pub commitment_scheme: CommitmentScheme,
-
-    /// Path to the verifier setup binary file
-    ///
-    /// Specifies the path to the verifier setup binary file required for verification.
-    /// Default path is `dynamic_dory.bin` in the `verifier_setups` directory.
-    #[arg(
-        long,
-        value_name = "VERIFIER_SETUP",
-        help = "Path to the verifier setup file",
-        default_value = "verifier_setups/dynamic_dory.bin"
-    )]
-    pub verifier_setup: String,
-}
-
-impl From<&QueryAndVerifySdkArgs> for (SxTClient, CommitmentScheme) {
-    fn from(args: &QueryAndVerifySdkArgs) -> Self {
-        (
-            SxTClient::new(
-                args.prover_root_url.clone(),
-                args.auth_root_url.clone(),
-                args.substrate_node_url.clone(),
-                args.sxt_api_key.clone(),
-                args.verifier_setup.clone(),
-            ),
-            args.commitment_scheme,
-        )
-    }
 }
